third_party/WebKit/LayoutTests/inspector/sources/debugger/debugger-no-pause-on-antibreakpoint-expected.txt

CONSOLE ERROR: line 9: Uncaught ReferenceError: unknown_var is not defined
CONSOLE ERROR: line 23: Uncaught TypeError: undefined is not a function
Tests that anti-breakpoint works.

Script source was shown.
Script execution paused.
Call stack:
    0) throwAnException (debugger-no-pause-on-antibreakpoint.html:9)
    1) handleClick (debugger-no-pause-on-antibreakpoint.html:14)
Script execution resumed.
Antibreakpoint must be set by now.
Script execution paused.
Call stack:
    0) handleClick2 (debugger-no-pause-on-antibreakpoint.html:23)
Script execution resumed.
